requirements.txt
asgiref==3.10.0
certifi==2025.10.5
channels==4.3.1
charset-normalizer==3.4.4
diff-match-patch==20241021
Django==5.2.7
django-cors-headers==4.9.0
django-countries==7.6.1
django-import-export==4.3.10
django-jazzmin==3.0.1
django-model-utils==5.0.0
django-notifications-hq==1.8.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
jsonfield==3.2.0
packaging==25.0
pi==0.1.2
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
routing==0.2.0
setuptools==68.0.0
sqlparse==0.5.3
swapper==1.4.0
tablib==3.9.0
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
utils==1.0.2
